incremental_build_mlx_mods.sh: Revert to config_and_prep
Commit 8a9cdbf7b848 ("Integrate changes from Amir") replaced the call to
`make dist-configs` in config_and_prep by a call to `make -C redhat
rh-configs`. I asked Amir and he did not remember exactly in which
situation the change was relevant, perhaps for RHEL8? Revert the change for
now in order to take advantage of the change in config_and_prep done in
the previous commit, bfbcf17 ("bashrc_addons.sh: Call dist-configs-arch
target to generate config files").